requirements.txt
tensorflow
opencv-python
pandas
pydicom
scikit-learn
numpy
pyyaml
